+++ UniBBS/TopicViewController.m
@@ -126,7 +126,7 @@
     if (buttonIndex == actionSheet.cancelButtonIndex) {
         return;
     }
-    int index = buttonIndex - actionSheet.firstOtherButtonIndex;
+    NSInteger index = buttonIndex - actionSheet.firstOtherButtonIndex;
     if (actionSheet.tag == ACTION_FROM_BAR_BUTTON) {
         // action sheet form bar buttom
         switch (index) {
